+++ src/web/components/locations/form.js
@@ -68,11 +68,33 @@ function LocationForm(values) {
     form.setComponents(formFactory.getComponents());
 
     form.setFormData({
+        type: 'create',
+        id: null,
         valid: false,
         result: null
     });
 
     form.addMethods({
+        disable: function() {
+            $(this.$el).find('input').attr('disabled', true);
+            $(this.$el).find('select').attr('disabled', true);
+        },
+
+        enable: function() {
+            $(this.$el).find('input').attr('disabled', false);
+            $(this.$el).find('select').attr('disabled', false);
+        },
+
+        changeType: function(type, id) {
+            this.$set('type', type);
+
+            if (type == 'create') {
+                this.$set('id', null);
+            } else {
+                this.$set('id', id);
+            }
+        },
+
         save: function () {
 
             this.clearErrors();
@@ -81,8 +103,6 @@ function LocationForm(values) {
 
                 if (valid) {
                     var that = this;
-
-
 
                     this.$parent.getLocationsApi().create(this.getValue(), function(err, data) {
 
@@ -103,6 +123,47 @@ function LocationForm(values) {
                         }
 
                         that.$parent.addLocation(data.result);
+                        that.clear();
+                        that.$parent.hideAddForm();
+                    });
+                }
+            });
+        },
+
+        update: function () {
+
+            this.clearErrors();
+
+            this.validate(function (err, valid) {
+
+                if (valid) {
+                    var that = this;
+
+                    var validData = this.getValue();
+
+                    var body = {
+                        id: this.id,
+                        response: validData.response
+                    };
+
+                    this.$parent.getLocationsApi().update(body, function(err, data) {
+
+                        if (err) {
+                            if (err.status
+                                && err.response
+                                && err.response.body
+                                && err.response.body.list
+                            ) {
+                                err.response.body.list.forEach(function(item) {
+                                    that.errors.push(item.message);
+                                });
+                            } else {
+                                console.log(err);
+                            }
+
+                            return;
+                        }
+
                         that.clear();
                         that.$parent.hideAddForm();
                     });