+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # todo
 
-- edit locations
 - delete locations
-
+- codemirror for response form field
 
 # changelog
 
 ## 2015-03-30
 - [be] saving locations into JSON-file, backup, restore locations
+- [api, web]  edit locations, only response
 
 ## 2015-03-28
 - [web] list locations and add location form